+++ tests/testthat/test_utils.R
@@ -34,6 +34,19 @@ test_that("getRowsFromPattern", {
                c(TRUE, FALSE, TRUE, TRUE))
 })
 
+test_that(".filterNA", {
+  m <- matrix(1:12, nrow=4)
+  m[c(2, 12)] <- NA_real_
+  expect_error(MSnbase:::.filterNA(1:10, 1, "110"), "must be a matrix")
+  expect_error(MSnbase:::.filterNA(m, pNA="110"), "must be numeric")
+  expect_error(MSnbase:::.filterNA(m, 1:10), "must be of length one")
+  expect_equal(MSnbase:::.filterNA(m), c(TRUE, FALSE, TRUE, FALSE))
+  expect_equal(MSnbase:::.filterNA(m, pNA=1), c(TRUE, TRUE, TRUE, TRUE))
+  expect_equal(MSnbase:::.filterNA(m, pNA=0.4), c(TRUE, TRUE, TRUE, TRUE))
+  expect_equal(MSnbase:::.filterNA(m, pNA=0.3), c(TRUE, FALSE, TRUE, FALSE))
+  expect_equal(MSnbase:::.filterNA(m, pattern="101"), c(TRUE, FALSE, TRUE, FALSE))
+})
+
 test_that("vec2ssv & ssv2vec", {
   numbers <- 1:3
   string1 <- "1;2;3"